ISSUE #4 Added taffydb and configured routes Philae and json request.
We have added taffydb to save data Philae. At this time the route is set to "get" because it receives no parameters.

We've added a json as persistent database on the server (db.json) file.

We set the path "/ viewer" to read the database and return a json string.

The "/ Philae" route has code to catch errors in the insert and update data.