ISSUE #17 Replace milligram with bootstrap
Use Bootstrap framework to build the frontend. Hopefully Cristian will
have time for some beatification in the future :)

Change also the build script, so that the index.html is part of the
source and the file is copied together with the rest of code and
libraries. This makes the design somewhat cleaner